Refactored Autocomplete out to it's own view
