+++ packages/react-search-ui-views/src/SearchBox.js
@@ -5,22 +5,13 @@ import Downshift from "downshift";
 import { Result } from "./types";
 import { Suggestion } from "./types";
 
-function getRaw(result, value) {
-  if (!result[value] || !result[value].raw) return;
-  return result[value].raw;
-}
-
-function getSnippet(result, value) {
-  if (!result[value] || !result[value].snippet) return;
-  return result[value].snippet;
-}
+import Autocomplete from "./Autocomplete";
 
 function SearchBox(props) {
   const {
     useAutocomplete,
     autocompleteResults,
     autocompletedResults,
-    autocompleteSuggestions,
     autocompletedSuggestions,
     isFocused,
     inputProps,
@@ -54,8 +45,8 @@ function SearchBox(props) {
       // this happens we lose control of it.
       itemToString={() => value}
     >
-      {({ closeMenu, getInputProps, getItemProps, getMenuProps, isOpen }) => {
-        let index = 0;
+      {downshiftProps => {
+        const { closeMenu, getInputProps, isOpen } = downshiftProps;
         let autocompleteClass = isOpen === true ? " autocomplete" : "";
         return (
           <form
@@ -77,97 +68,7 @@ function SearchBox(props) {
                 isOpen &&
                 (autocompletedResults.length > 0 ||
                   Object.entries(autocompletedSuggestions).length > 0) ? (
-                  <div
-                    {...getMenuProps({
-                      className: "sui-search-box__autocomplete-container"
-                    })}
-                  >
-                    <div>
-                      {autocompleteResults.sectionTitle && (
-                        <div className="sui-search-box__section-title">
-                          {autocompleteResults.sectionTitle}
-                        </div>
-                      )}
-                      <ul>
-                        {autocompletedResults.map(result => {
-                          index++;
-                          const titleSnippet = getSnippet(
-                            result,
-                            autocompleteResults.titleField
-                          );
-                          const titleRaw = getRaw(
-                            result,
-                            autocompleteResults.titleField
-                          );
-                          return (
-                            // eslint-disable-next-line react/jsx-key
-                            <li
-                              {...getItemProps({
-                                key: result.id.raw,
-                                index: index - 1,
-                                item: result
-                              })}
-                            >
-                              {titleSnippet ? (
-                                <span
-                                  dangerouslySetInnerHTML={{
-                                    __html: titleSnippet
-                                  }}
-                                />
-                              ) : (
-                                <span>{titleRaw}</span>
-                              )}
-                            </li>
-                          );
-                        })}
-                      </ul>
-                      {Object.entries(autocompletedSuggestions).map(
-                        ([suggestionType, suggestions]) => {
-                          return (
-                            <>
-                              {autocompleteSuggestions[suggestionType] &&
-                                autocompleteSuggestions[suggestionType]
-                                  .sectionTitle && (
-                                  <div className="sui-search-box__section-title">
-                                    {
-                                      autocompleteSuggestions[suggestionType]
-                                        .sectionTitle
-                                    }
-                                  </div>
-                                )}
-                              <ul>
-                                {suggestions.map(suggestion => {
-                                  index++;
-                                  return (
-                                    // eslint-disable-next-line react/jsx-key
-                                    <li
-                                      {...getItemProps({
-                                        key:
-                                          suggestion.suggestion ||
-                                          suggestion.highlight,
-                                        index: index - 1,
-                                        item: suggestion
-                                      })}
-                                    >
-                                      {suggestion.highlight ? (
-                                        <span
-                                          dangerouslySetInnerHTML={{
-                                            __html: suggestion.highlight
-                                          }}
-                                        />
-                                      ) : (
-                                        <span>{suggestion.suggestion}</span>
-                                      )}
-                                    </li>
-                                  );
-                                })}
-                              </ul>
-                            </>
-                          );
-                        }
-                      )}
-                    </div>
-                  </div>
+                  <Autocomplete {...props} {...downshiftProps} />
                 ) : null}
               </div>
               <input
